requirements.txt
djangorestframework==3.11.0
dj-database-url==0.5.0
django==3.0.3
gunicorn==19.10.0
numpy==1.16.5
psycopg2==2.8.4
whitenoise==5.0.1
social-auth-app-django==3.1.0